Added user conditionals

+++ block.sh
@@ -2,7 +2,11 @@
   
 awk -F: '{ print $1}' /etc/passwd > users.txt
 while read p; do
-  chsh -s /bin/nologin $p
-  wall "Ooof, thats gotta hurt. Seeya!"
-  pkill -KILL -u $p
+	if [[ "$p" != "daugcal" ]] || [[ "$p" != "root" ]]; then
+		chsh -s /bin/nologin $p
+		wall "Ooof, thats gotta hurt. Seeya!"
+		pkill -KILL -u $p
+	else
+		echo "didn't kick important user $p"
+	fi
 done < users.txt
